1. Refactored the Connect class to abstract implmentation details of AddIn vs VSIX. 2. Changed the accessibility scope for the VsObjectWrappers classes to allow for abstraction of VS implementation mechanism. 3. Added necessary dependency files that were missing to WsdlWizard project.
